refactor: separate data protocols in folders
