+++ src/data/protocols/cryptography/index.ts
@@ -1,0 +1,1 @@
+export * from './encrypter'
