+++ src/data/protocols/db/error-log-repository.ts
