+++ src/data/protocols/db/index.ts
@@ -1,0 +1,3 @@
+export * from './add-account-repository'
+export * from './error-log-repository'
+export * from './load-account-by-email-repository'
