+++ src/data/protocols/index.ts
@@ -1,4 +1,2 @@
-export * from './encrypter'
-export * from './add-account-repository'
-export * from './error-log-repository'
-export * from './load-account-by-email-repository'
+export * from './cryptography'
+export * from './db'
